A little code sweeping.
Removed createListModel and createItem and replaced by with putListModel and putItem.

add was renamed to addOrUpdate

Finally, I change the remove and addOrUpdate functions in such a way so that they are now responsible for deleting and putting items to the database as well as creating links to them.

Love to hear thoughts.